backend/requirements.txt
Django==6.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1
python-dotenv==1.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
Pillow==10.1.0
